+++ CHANGES.md
@@ -4,6 +4,10 @@ moodle-local_boostnavigation
 Changes
 -------
 
+### Unreleased
+
+* 2018-07-23 - Remove deprecated strings file as the strings were fully depreated before.
+
 ### v3.5-r2
 
 * 2018-07-19 - Adapt the mechanism to de-activate collapsible custom parent nodes.